[WIP] [Request feedback] Adding support for nullable and default valuers in DDL creation
As a developer efficiency improvement, we would like to introduce the ability to specify if a
column is nullable or not, and add the ability to set a default value, for DDLs.  In this shot
this is passed through to the JDBC module and nothing else, however it can be introduced into
other modules as well.