Adding support for NOT NULL in DDL statements
This change enables NOT NULL in the parser and adds a flag in the
metadata to check whether or not a connector supports this new
functionality.  If a connector doesn't support NOT NULL semantics in
columns, yet this is supplied, then a semantic exception is thrown
indicating to the user that this functionality is not supported in their
catalog.